Fix hang when cmgr/pmgr initialize early
This fixes what seems to be a Catalina-specific issue.

Pre-Catalina: The DidUpdateState delegate callbacks get called *after*
the delegate properties gets assigned.

Catalina: DidUpdateState gets called immediately upon construction of
the manager objects (CBCentralManager and CBPeripheralManager).

Prior to this commit, the code assumed it could assign the delegate
properties and then listen for the transition to the powered-on state.
This works for older versions of macOS, but with Catalina the code
blocks forever, waiting for an event that has already occurred.

The fix is to use the following procedure:

(for each manager object):
1. Assign the delegate.
2. Check the state.  If the state is not "unknown" then the procedure is
done.
3. Else, wait up to one second for a state change event.